+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - New `GlimpseNetCDF()` function that is an alias for `ReadNetCDF(out = "vars")` but now prints a human readable summary of the contents of the NetCDF file.
 
+- `geom_streamline()` now uses 4th order Runge-Kutta insteaf of plain old Euler. It also draws arrows in the middle of the streamline. 
+
 ## Bugfixes
 
 - `ReadNetCDF()` result will print correctly.